requirements.txt
streamlit
PyMuPDF
pandas
openpyxl
python-docx
python-pptx
faiss-cpu
openai
numpy
azure-storage-blob
